refactor(database): re-introduce patch/replace functions refactor(database): remove authz code from crud handlers refactor(database): forward authz params to queries for crud/custom endpoints

+++ modules/database/src/handlers/CustomEndpoints/customEndpoint.handler.ts
@@ -32,6 +32,7 @@ export class CustomEndpointHandler {
     const path = call.request.path.split('/')[3];
     const endpoint: ICustomEndpoint = CustomEndpointHandler.routeControllers[path];
     const params = call.request.params;
+    const scope = call.request.queryParams.scope;
     let searchQuery: Indexable = {};
     let createString = '';
 
@@ -130,12 +131,17 @@ export class CustomEndpointHandler {
           .model.findMany(searchQuery, {
             skip: params['skip'],
             limit: params['limit'],
+            scope,
+            userId: call.request.context.user?._id,
             sort,
             populate: params['populate'],
           });
         const countPromise = this.database
           .getSchemaModel(endpoint.selectedSchemaName)
-          .model.countDocuments(searchQuery);
+          .model.countDocuments(searchQuery, {
+            scope,
+            userId: call.request.context.user?._id,
+          });
         promise = Promise.all([documentsPromise, countPromise]);
       } else {
         promise = this.database
@@ -143,24 +149,32 @@ export class CustomEndpointHandler {
           .model.findMany(searchQuery, {
             sort,
             populate: params['populate'],
+            scope,
+            userId: call.request.context.user?._id,
           });
       }
     } else if (endpoint.operation === OperationsEnum.POST) {
       promise = this.database
         .getSchemaModel(endpoint.selectedSchemaName)
-        .model.create(createObj);
+        .model.create(createObj, { scope, userId: call.request.context.user?._id });
     } else if (endpoint.operation === OperationsEnum.PUT) {
       promise = this.database
         .getSchemaModel(endpoint.selectedSchemaName)
-        .model.updateMany(searchQuery, createObj);
+        .model.updateMany(searchQuery, createObj, {
+          scope,
+          userId: call.request.context.user?._id,
+        });
     } else if (endpoint.operation === OperationsEnum.DELETE) {
       promise = this.database
         .getSchemaModel(endpoint.selectedSchemaName)
-        .model.deleteMany(searchQuery);
+        .model.deleteMany(searchQuery, { scope, userId: call.request.context.user?._id });
     } else if (endpoint.operation === OperationsEnum.PATCH) {
       promise = this.database
         .getSchemaModel(endpoint.selectedSchemaName)
-        .model.updateMany(searchQuery, createObj);
+        .model.updateMany(searchQuery, createObj, {
+          scope,
+          userId: call.request.context.user?._id,
+        });
     } else {
       process.exit(-1);
     }
